Add admin and class log seawrch to a separate page #2983
