+++ static/js/index.ts
@@ -2,6 +2,7 @@ export * from './modal';
 export * from './app';
 export * from './auth';
 export * from './statistics';
+export * from './logs';
 export * from './tutorial';
 export * from './quiz';
 export * from './parsons';
