+++ templates/admin/admin-stats.html
@@ -1,6 +1,6 @@
 {% extends "auth.html" %}
 
-{% from "stats-shared.html" import insert_log_search, insert_accordion_tab %}
+{% from "stats-shared.html" import insert_accordion_tab %}
 
 {% block main %}
 <div class="flex flex-row mb-4 border-b">
@@ -158,10 +158,6 @@
           </div>
         </div>
       </div>
-    {% endcall %}
-
-    {% call insert_accordion_tab('tab-multi-seven', 'Search program logs') %}
-        {{ insert_log_search() }}
     {% endcall %}
 
   </div>
