+++ templates/admin/admin.html
@@ -14,6 +14,7 @@
         <button class="green-btn text-xl m-2 p-4 section-button"  onclick="window.open('/admin/adventures', '_self')">Adventures overview</button>
         <button class="green-btn text-xl m-2 p-4 section-button"  onclick="window.open('/admin/achievements', '_self')">Achievements overview</button>
         <button class="green-btn text-xl m-2 p-4 section-button"  onclick="window.open('/admin/stats', '_self')">Statistics</button>
+        <button class="green-btn text-xl m-2 p-4 section-button"  onclick="window.open('/admin/logs', '_self')">Logs</button>
     </div>
 </div>
 {% endblock %}
