+++ website/admin.py
@@ -129,6 +129,14 @@ def routes(app, database):
             return utils.error_page(error=403, ui_message=gettext('unauthorized'))
         return render_template('admin/admin-stats.html', page_title=gettext('title_admin'))
 
+    @app.route('/admin/logs', methods=['GET'])
+    @requires_login
+    def get_admin_logs_page(user):
+        if not is_admin(user):
+            return utils.error_page(error=403, ui_message=gettext('unauthorized'))
+        return render_template('admin/admin-logs.html', page_title=gettext('title_admin'))
+
+
     @app.route('/admin/achievements', methods=['GET'])
     @requires_login
     def get_admin_achievements_page(user):
